Data completed - Format Incomplete
The background image is from Chrono Cross videogame from PlayStation and
the stage is outside roof of Viper Manor.

html{}

Just like my previous project, I set up font-size in html as 62.5%,
Default it is still 16px, but I can work with REM units. So, 1rem = 10px
because so then I can work under this: 10px= 1rem.

According to eedrah from
https://stackoverflow.com/questions/7714543/font-size62-5-vs-font-size10px

If we just set up 10px, this method will override the user's
accessibility options.

body{} I set up margin 0 in order to stick the image to the screen. The
x scroll should not be seen so overflow-x is hidden.

Divs background-chronocross-img and overlay color, they create the
overlay effect. If I am not wrong, it takes way too time time to load,
but I haven't checked the answer code in freecodecamp, I wanted to do it
from scratch with the notes I took and some internet research. I'll
check it afterwards in order to solve it with the most appropiate
solution.
The image size is cover so it sticks to the window.

Main is a centered block with a width of 500px, the rest of the
characteristics are display block and margin 0.

Header, I set up title and a description. Nothing very relevant.

Form-box class, I wanted to change the whole content font alignment and
its padding. So, I did so in CSS.

For the rest of the form fields, it is still on process. I am just
experimenting at this point. I haven't defined classes for this as a lot
of specific features must be repeated. So I applied the format in order
to see what looked best but in terms of good practice, it has not been
developed properly. I'll review this one more time in order to put some
insight about it and next time I hope I can finish the project.

There will be two versions. One version, my own path with all the
imperfections all the way. Then, the freecodecamp match up version.

Thanks for reading!